+++ R/zzz.R
@@ -1,8 +1,7 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.aion <- list(
-    aion.precision = 1,
-    aion.calendar = calendar("CE")
+    aion.calendar = NULL
   )
   toset <- !(names(op.aion) %in% names(op))
   if(any(toset)) options(op.aion[toset])