+++ inst/tinytest/test_series.R
@@ -1,5 +1,4 @@
 Sys.setenv(LANGUAGE = "en") # Force locale
-options(aion.calendar = calendar("CE"))
 
 # Create from vector ===========================================================
 x <- rnorm(91)